feat: linux set default printer and remove printer
